pyproject.toml
[tool.poetry]
name = "vk-bot"
version = "0.1.0"
description = "VK store bot"
authors = ["Vjacheslav <goplit2010.konovalov@Yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
vk-api = "^11.9.7"
python-dotenv = "^0.19.2"
PyMySQL = "^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
